test(e2e): consolidate provider operation helpers
