+++ tests/e2e/provider_operation_google_docs_cases.py
@@ -2,7 +2,7 @@
 
 import json
 
-from provider_operation_google_common import google_json
+from emulate_provider import google_json
 from provider_operation_types import ProviderOperationCase
 
 DOCUMENT_ID = "doc_reborn_strategy"
